Move Name file to Client package
