+++ src/main/java/seedu/address/model/person/Client.java
@@ -2,6 +2,7 @@ package seedu.address.model.person;
 
 import java.util.Set;
 
+import seedu.address.model.client.Name;
 import seedu.address.model.project.Project;
 import seedu.address.model.tag.Tag;
