+++ src/main/java/seedu/address/model/person/Person.java
@@ -7,6 +7,7 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 
+import seedu.address.model.client.Name;
 import seedu.address.model.tag.Tag;
 
 /**
